resolved security issue

+++ sb-ecom/src/main/java/com/ecommerce/project/payload/CategoryResponse.java
@@ -10,10 +10,10 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CategoryResponse {
- private List<CategoryDTO> content;
- private Integer pageNumber;
- private Integer pageSize;
- private Long totalElements;
- private Long totalPages;
- private boolean lastPage;
+    private List<CategoryDTO> content;
+    private Integer pageNumber;
+    private Integer pageSize;
+    private Long totalElements;
+    private Integer totalPages;
+    private boolean lastPage;
 }
