+++ sb-ecom/src/main/java/com/ecommerce/project/payload/ProductResponse.java
@@ -6,16 +6,14 @@ import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductResponse {
- private List<ProductDTO> content;
- private Integer pageNumber ;
- private Integer pageSize ;
- private Integer totalElements ;
- private Integer totalPages ;
- private boolean lastPage  ;
+    private List<ProductDTO> content;
+    private Integer pageNumber;
+    private Integer pageSize;
+    private Long totalElements;
+    private Integer totalPages;
+    private boolean lastPage;
 }
-
